libstagefright: adjust audio seek for mkv container
The original seeking method can only deal with the case
that the SimpleBlocks of both audio and video are packed
in the head of a cluster. The testing clip is quite
special, as the key frame for video is located at the end
of a cluster, at 10000ms, while the first SimpleBlock of
audio in this cluster starts at 5013ms. So video is stopped
untill audio catches up.

After modification, audio will advance to the nearest point
as video has seeked to.

Change-Id: Ic03abea1ebb8ef630ae9000dd904c2ce3986d5af
Author: Dan Liang <dan.liang@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 57484